minor refactor: create file for member class
